Testing"

+++ globals.h
@@ -3,9 +3,11 @@
 
 // QUIZ_QUESTIONS_COUNT should be 5 or its multipliers (10, 15, 20)
 const int QUIZ_QUESTIONS_COUNT = 5;
-const double MCQCount = 0.4 * QUIZ_QUESTIONS_COUNT;
-const double TFCount = 0.4 * QUIZ_QUESTIONS_COUNT;
-const double CompleteCount = 0.2 * QUIZ_QUESTIONS_COUNT;
+
+const int MCQCount = (int) (0.4 * QUIZ_QUESTIONS_COUNT);
+const int TFCount = (int) (0.4 * QUIZ_QUESTIONS_COUNT);
+const int CompleteCount = (int) (0.2 * QUIZ_QUESTIONS_COUNT);
+
 const double HighestPossibleScore = (MCQCount * 2) + TFCount + (CompleteCount * 3);
 int POOL_QUESTIONS_COUNT = 0;
